python/pyproject.toml
[project]
name = "infinitesense"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pyqt6>=6.9.1",
    "pyqtgraph>=0.13.7",
    "pyserial>=3.5",
    "pyzmq>=27.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.18.1",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]
